impltest

+++ src/main/java/org/iii/ee100/sample/lala/CalculateImpl.java
@@ -23,7 +23,7 @@ public class CalculateImpl implements Calculate {
 
 	public double div(int a, int b) {
 		
-		return a/b;
+		return (double)a/(double) b;
 	}
